hackrx-flask/requirements.txt
pandas==1.5.3
numpy==1.22.4
matplotlib==3.7.1
nltk==3.8.1
scikit-learn==1.2.2
sklearn-pandas==2.2.0

pickle
re
warnings
